requirements.txt
fonttools>=4.38.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.20.0
scipy>=1.10.0